Move builder classes into their own package namespace
