+++ odata-client-msgraph-beta/src/main/java/com/github/davidmoten/msgraph/beta/MsGraph.java
@@ -1,7 +1,6 @@
 package com.github.davidmoten.msgraph.beta;
 
-import com.github.davidmoten.msgraph.MsGraphClientBuilder;
-import com.github.davidmoten.msgraph.MsGraphClientBuilder.Builder;
+import com.github.davidmoten.msgraph.builder.MsGraphClientBuilder;
 
 import odata.msgraph.client.beta.container.GraphService;
 
@@ -13,7 +12,7 @@ public final class MsGraph {
         // prevent instantiation
     }
 
-    public static Builder<GraphService> tenantName(String tenantName) {
+    public static MsGraphClientBuilder.Builder<GraphService> tenantName(String tenantName) {
         return new MsGraphClientBuilder<GraphService>(MSGRAPH_BETA_BASE_URL,
                 context -> new GraphService(context)).tenantName(tenantName);
     }
