+++ odata-client-msgraph/src/main/java/com/github/davidmoten/msgraph/MsGraph.java
@@ -1,6 +1,6 @@
 package com.github.davidmoten.msgraph;
 
-import com.github.davidmoten.msgraph.MsGraphClientBuilder.Builder;
+import com.github.davidmoten.msgraph.builder.MsGraphClientBuilder;
 
 import odata.msgraph.client.container.GraphService;
 
@@ -12,7 +12,7 @@ public final class MsGraph {
         // prevent instantiation
     }
 
-    public static Builder<GraphService> tenantName(String tenantName) {
+    public static MsGraphClientBuilder.Builder<GraphService> tenantName(String tenantName) {
         return new MsGraphClientBuilder<GraphService> //
         (MSGRAPH_1_0_BASE_URL, //
                 context -> new GraphService(context)).tenantName(tenantName);
